requirements.txt
Django
psycopg2
django-hvad
